Move ics20 in application
