+++ modules/src/application/ics20_fungible_token_transfer/mod.rs
@@ -1,2 +1,4 @@
 //! ICS 20: IBC Transfer implementation
+pub mod error;
 pub mod events;
+pub mod msgs;
