+++ modules/src/application/mod.rs
@@ -1,2 +1,1 @@
-pub mod error;
-pub mod msgs;
+pub mod ics20_fungible_token_transfer;
