+++ modules/src/lib.rs
@@ -17,10 +17,11 @@
 //! - ICS 04: Channel
 //! - ICS 07: Tendermint Client
 //! - ICS 18: Basic relayer functions
-//! - ICS 20: Fungible Token
 //! - ICS 23: Vector Commitment Scheme
 //! - ICS 24: Host Requirements
 //! - ICS 26: Routing
+//! - Applications:
+//!    - ICS 20: Fungible Token Transfer
 pub mod address;
 pub mod application;
 pub mod events;
@@ -30,7 +31,6 @@ pub mod ics03_connection;
 pub mod ics04_channel;
 pub mod ics07_tendermint;
 pub mod ics18_relayer;
-pub mod ics20_fungible_token_transfer;
 pub mod ics23_commitment;
 pub mod ics24_host;
 pub mod ics26_routing;
